backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.10.3
pydantic[email]==2.10.3
httpx==0.28.1
requests==2.32.3
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.6.1

# Logging
structlog==24.4.0

# APM / Error Tracking
sentry-sdk[fastapi]==2.19.2

# Monitoring
prometheus-client>=0.17.0

# Encryption (for backup encryption)
cryptography>=41.0.0
